+++ src/datumaro/components/environment.py
@@ -74,10 +74,9 @@ class Environment:
     def __init__(self):
         from datumaro.components.dataset_base import (
             DatasetBase,
-            ItemTransform,
             SubsetBase,
-            Transform,
         )
+        from datumaro.components.transformer import ItemTransform, Transform
         from datumaro.components.importer import Importer
         from datumaro.components.exporter import Exporter
         from datumaro.components.generator import DatasetGenerator
