+++ src/datumaro/components/launcher.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from datumaro.components.annotation import AnnotationType, LabelCategories
 from datumaro.components.cli_plugin import CliPlugin
-from datumaro.components.dataset_base import Transform
+from datumaro.components.transformer import Transform
 from datumaro.util import take_by
