+++ src/datumaro/plugins/sampler/random_sampler.py
@@ -9,7 +9,8 @@ from typing import List, Mapping, Optional, Tuple
 
 from datumaro.components.annotation import AnnotationType
 from datumaro.components.cli_plugin import CliPlugin
-from datumaro.components.dataset_base import DatasetItem, IDataset, Transform
+from datumaro.components.dataset_base import DatasetItem, IDataset
+from datumaro.components.transformer import Transform
 from datumaro.util import cast
