+++ src/datumaro/plugins/sampler/relevancy_sampler.py
@@ -8,7 +8,8 @@ from typing import Optional, Union
 import pandas as pd
 
 from datumaro.components.cli_plugin import CliPlugin
-from datumaro.components.dataset_base import IDataset, Transform
+from datumaro.components.dataset_base import IDataset
+from datumaro.components.transformer import Transform
 from datumaro.util import parse_str_enum_value
 
 from .algorithm.algorithm import Algorithm, SamplingMethod
