+++ src/datumaro/plugins/splitter.py
@@ -11,7 +11,7 @@ import numpy as np
 
 from datumaro.components.annotation import AnnotationType
 from datumaro.components.cli_plugin import CliPlugin
-from datumaro.components.dataset_base import Transform
+from datumaro.components.transformer import Transform
 from datumaro.util import cast
 from datumaro.util.definitions import DEFAULT_SUBSET_NAME
