+++ tests/unit/test_project.py
@@ -9,7 +9,8 @@ import numpy as np
 from datumaro.components.annotation import Bbox, Label
 from datumaro.components.config_model import Model, Source
 from datumaro.components.dataset import DEFAULT_FORMAT, Dataset
-from datumaro.components.dataset_base import DatasetBase, DatasetItem, ItemTransform
+from datumaro.components.dataset_base import DatasetBase, DatasetItem
+from datumaro.components.transformer import ItemTransform
 from datumaro.components.errors import (
     DatasetMergeError,
     EmptyCommitError,
